# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Sign-in API
  version: 0.1.0
  description: Account registration, browser session authentication and profile management.
  license:
    name: Proprietary
    url: https://github.com/FirewolfO/sigin
servers:
  - url: /api/v1
paths:
  /auth/csrf:
    get:
      operationId: initializeCsrf
      summary: Initialize the browser CSRF cookie
      security: []
      responses:
        '200':
          description: CSRF cookie initialized
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CsrfMetadata'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/register:
    post:
      operationId: register
      summary: Register an account and create a session
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      operationId: login
      summary: Log in with username, email or phone
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the current session user
      security:
        - sessionCookie: []
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      operationId: logout
      summary: Invalidate the current session
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Session invalidated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [loggedOut]
                        properties:
                          loggedOut:
                            type: boolean
                            const: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /account/profile:
    put:
      operationId: updateProfile
      summary: Update the current account profile
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: CLOUD_SESSION
  parameters:
    CsrfHeader:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: Value from the XSRF-TOKEN cookie returned by /auth/csrf.
      schema:
        type: string
  schemas:
    Envelope:
      type: object
      required: [code, message, data, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        data: {}
        requestId:
          type: string
    CsrfMetadata:
      type: object
      required: [headerName, parameterName]
      properties:
        headerName:
          type: string
          const: X-XSRF-TOKEN
        parameterName:
          type: string
          const: _csrf
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9_.-]{2,31}$'
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
        displayName:
          type: [string, 'null']
          maxLength: 64
        email:
          type: [string, 'null']
          format: email
          maxLength: 254
        phone:
          type: [string, 'null']
          pattern: '^\+[1-9]\d{7,14}$'
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          maxLength: 254
        password:
          type: string
          format: password
          maxLength: 72
    UpdateProfileRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 64
        email:
          type: [string, 'null']
          format: email
          maxLength: 254
        phone:
          type: [string, 'null']
          pattern: '^\+[1-9]\d{7,14}$'
        avatarUrl:
          type: [string, 'null']
          format: uri
          maxLength: 1000
    User:
      type: object
      required: [id, username, displayName, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        email:
          type: [string, 'null']
          format: email
        phone:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
          format: uri
        lastLoginAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Error:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: 'null'
  responses:
    UserResponse:
      description: User response
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing session or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Missing or invalid CSRF token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Username, email or phone is already in use
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'